obs-websocket/requirements.txt
obs-websocket-py
